Move os to util
