+++ libopenage/crossplatform/CMakeLists.txt
@@ -1,3 +1,2 @@
 add_sources(libopenage
-	os.cpp
 )
