+++ libopenage/util/os.h
